feat(ArticleController, article concered View) : createForm으로 이동할 때 nickname 추가, delete, update는 권한 있는 사용자만 가능하게 한다.

+++ src/main/java/com/y/java_board/controller/ArticleController.java
@@ -6,30 +6,35 @@ import com.y.java_board.dto.ArticleDto;
 import com.y.java_board.dto.CommentDto;
 import com.y.java_board.service.ArticleService;
 import com.y.java_board.service.CommentService;
+import com.y.java_board.service.UserService;
+import jakarta.servlet.http.HttpServletRequest;
+import lombok.AllArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
+import org.springframework.web.servlet.support.RequestContextUtils;
 
+import java.security.Principal;
 import java.util.List;
-import java.util.NoSuchElementException;
+import java.util.Map;
 import java.util.Optional;
 
-@Controller()
+@Controller
+@AllArgsConstructor
 public class ArticleController {
 
     private final Logger logger = LoggerFactory.getLogger(getClass());
     private final ArticleService articleService;
     private final CommentService commentService;
-
-    public ArticleController(ArticleService articleService, CommentService commentService){
-        this.articleService = articleService;
-        this.commentService = commentService;
-    }
+    private final UserService userService;
 
     @GetMapping("/articles")
     public String articles(Model model){
@@ -39,8 +44,9 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/new")
-    public String showCreateForm(Model model){
-        model.addAttribute("articleDto", new ArticleDto("","",""));
+    public String showCreateForm(Model model, @AuthenticationPrincipal UserDetails userDetails){
+        String nickname = userService.getNicknameByEmail(userDetails.getUsername());
+        model.addAttribute("articleDto", new ArticleDto("","", nickname));
         return "article/create";
     }
 
@@ -55,13 +61,18 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/{id}")
-    public String showArticle(@PathVariable("id") long id, Model model){
+    public String showArticle(@PathVariable("id") long id, Model model, Principal principal, HttpServletRequest request){
         Optional<Article> articleOptional = articleService.findOne(id);
+        Map<String, ?> inputFlashMap = RequestContextUtils.getInputFlashMap(request);
         if(articleOptional.isPresent()){
             List<Comment> comments = commentService.findCommentsByArticleId(id);
             model.addAttribute("article",articleOptional.get());
-            model.addAttribute("commentDto", new CommentDto("", "", id));
+            String nickname = userService.getNicknameByEmail(principal.getName());
+            model.addAttribute("commentDto", new CommentDto(nickname, "", id));
             model.addAttribute("comments", comments);
+            if(inputFlashMap != null){
+                model.addAttribute("auth", inputFlashMap.get("auth"));
+            }
             return "/article/detail";
         }
         // TODO : 해당 id 가 없다는 안내가 필요할 것 같다.
@@ -69,17 +80,27 @@ public class ArticleController {
     }
 
     @GetMapping("/articles/delete/{id}")
-    public String deleteArticle(@PathVariable("id") long id, Model model){
+    public String deleteArticle(@PathVariable("id") long id, Model model, Principal principal, RedirectAttributes redirectAttributes){
         Article article = articleService.findOne(id)
                 .orElseThrow(()-> new IllegalArgumentException("Invalid article Id : " + id));
+        String email = userService.getEmailByNickname(article.getWriter());
+        if(!principal.getName().equals(email)){
+            redirectAttributes.addFlashAttribute("auth", "삭제 권한이 없습니다.");
+            return "redirect:/articles/{id}";
+        }
         articleService.deleteOne(id);
         return "redirect:/articles";
     }
 
     @GetMapping("/articles/update/{id}")
-    public String showUpdateForm(@PathVariable long id, Model model){
+    public String showUpdateForm(@PathVariable long id, Model model, Principal principal, RedirectAttributes redirectAttributes){
         Article article = articleService.findOne(id)
                 .orElseThrow(()-> new IllegalArgumentException("Invalid article Id: "+ id));
+        String email = userService.getEmailByNickname(article.getWriter());
+        if(!principal.getName().equals(email)){
+            redirectAttributes.addFlashAttribute("auth", "수정 권한이 없습니다.");
+            return "redirect:/articles/{id}";
+        }
         model.addAttribute("articleDto", new ArticleDto(article.getTitle(),article.getContent(), article.getWriter()));
         return "/article/update";
     }
